0.10.2.2
-Added Phoenix PowerUp
-Added Rare, Ultra Rare and legendary PowerUp pick ups
-Added Bubble Shield PowerUp
-Added RemovePowerUp method to PowerUpAdder
-Improved ChainLighning PowerUp Effect
-Added Coin Bundle
-Orange Enemy shield now scales with it's health
-Green enemies will now stop moving and rotating when shooting
-Improved HealthGradient script

+++ Assets/C# Scripts/Item.cs
@@ -9,7 +9,7 @@ public class Item : MonoBehaviour
     [System.Serializable]
     public struct Loot
     {
-        public int Weight;
+        public float Weight;
         public PowerUp Item;
     }
